atom: Calls to size_ptr2array_ref are no longer necessary.

+++ include/ad_tensor/dev/agraph.hpp
@@ -75,7 +75,7 @@ public:
     vector< vector<int64_t> >    m_dom_shapes;
     vector<op_enum_t>            m_op_seq;
     vector<size_t>               m_arg_start;
-    vector<size_t>               m_arg_value;
+    vector<int64_t>              m_arg_value;
     vector<ad_type_t>            m_arg_type;
     //
     // default constructor
